+++ lib/app/modules/policies/view_policy.dart
@@ -1,5 +1,6 @@
 import 'package:flutter/material.dart';
-import 'package:money/app/core/helpers/color_helper.dart';
+import 'package:flutter_markdown/flutter_markdown.dart';
+import 'package:money/app/data/models/constants.dart';
 
 class PolicyScreen extends StatelessWidget {
   const PolicyScreen({
@@ -9,29 +10,36 @@ class PolicyScreen extends StatelessWidget {
   @override
   Widget build(BuildContext context) {
     return Center(
-      child: SizedBox(
-        width: 400,
-        child: SelectableText(
-          style: getTextTheme(context).bodySmall,
-          '''
-      Privacy Policy for MyMoney App
-      
-      1. No Information Collected:
-      MyMoney does not collect any personal information from its users. We do not require users to provide any personal data such as name, email address, or any other identifying information.
-      
-      2. Information Usage:
-      Since we do not collect any personal information, we do not use or share any information about our users.
-      
-      3. No Data Logged:
-      MyMoney does not log any data from its users.
-      
-      4. Contact Us:
-      If you have any questions or suggestions about our Privacy Policy, do not hesitate to contact us at questions@vteam.com.
-      
-      By using MyMoney, you signify your acceptance of this Privacy Policy. If you do not agree to this policy, please do not use our application. Your continued use of the application following the posting of changes to this policy will be deemed your acceptance of those changes.</p>
-      ''',
+      child: Padding(
+        padding: const EdgeInsets.all(SizeForPadding.large),
+        child: Markdown(
+          data: getMDContext(),
+          selectable: true,
         ),
       ),
     );
+  }
+
+  String getMDContext() {
+    return '''
+# Privacy Policy for MyMoney App
+
+## 1. No Information Collected:
+MyMoney does not collect any personal information from its users. We do not require users to provide any personal data such as name, email address, or any other identifying information.
+
+## 2. Information Usage:
+Since we do not collect any personal information, we do not use or share any information about our users.
+
+## 3. No Data Logged:
+MyMoney does not log any data from its users.
+
+## 4. Contact Us:
+If you have any questions or suggestions about our Privacy Policy, do not hesitate to contact us at questions@vteam.com.
+
+_________________
+
+
+By using MyMoney, you signify your acceptance of this Privacy Policy. If you do not agree to this policy, please do not use our application. Your continued use of the application following the posting of changes to this policy will be deemed your acceptance of those changes.
+''';
   }
 }